Move HTTP-request related stuff to api.py
